client.shared.software_manager: Check if yum module is present
When using the software_manager component on a Red Hat like distro,
if the yum module for Python is not installed, log the information
as an error message and use the basic support from rpm/yum.

If the module yum is not installed (for example, in a virtualenv),
the software_manager component will not traceback when creating
an instance of YumBackend and only the method provides() will be disabled.

Also, Fedora versions starting with 22 don't have yum anymore,
as it has been replaced by DNF.

Signed-off-by: Ruda Moura <rmoura@redhat.com>