os_dep: refactor into common functions
Rewrite commands, libraries and headers to use common searching functions.

The lib searcher was re-implemented to match as close as possible the C
ldconfig implementation.  The C source code was translated line-by-line
as close as possible into python, so we preserve the exact same
behaviour with regards to recurrsion.

Unittests included to check all the interesting corner cases.

Also try to match `which` behaviour as close as possible.

Disabled pytest specific unittests.

Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>