scheduler: fix corner case with metahost scheduling on reserved hosts
Consider having a host reserved in Autotest with label 'L'. Run a
job-by-label (metahost) for 1 label-'L' host. The current scheduler will
not exclude your reserved host from the set of possible hosts to run the
metahost job on, because it only checks that your user has permission to
run on candidate hosts, not if those candidate hosts are already in-use.
I'm assuming that reservations in Autotest are specifically intended to
indicate a host is "in-use". I think this is the principle of least
surprise for reservations. With the changes in this commit, candidate
hosts' ACLs are checked against the "Everyone" ACL for set-identity and
excluded if there is an ACL in place.

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>