reservations: fix bug when multiple hosts reserved by same user
I noticed that if one reserves a system from the web UI, then reserves
a second system, the first reservation disappears. This appears to be
because reservation creation's first step is to remove all hosts from
the user_acl object in the database via user_acl.hosts.clear(), which
removes any entries for the corresponding 'name' in the afe_acl_groups
table from the afe_acl_groups_hosts table. This is pretty clearly not
the intent of the create() logic, which should not affect any of a
given user's existing reservations.

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>