scheduler: when aborting jobs due to timeout, disable netboot
Currently, if a job times, the host is left in the to-be-installed
state.  However, this can result (if many hosts fail to install), in
many hosts being left in a to-be-installed state without much
indiciation. It can also lead to unexpected user conditions if someone
reboots that system to use it's prior existing OS. It is better to leave
the system in as-close to an unaltered state as possible.

Signed-off-by: Nishanth Aravamudan <nacc@linux.vnet.ibm.com>