pylint: Ignore false-negatives for unsubscriptable values
The latest pylint is more picky about E1136 and produces couple of
false-negatives. Some are handled by code (check if is None), some
are caused by urlparse, which on latest versions returns custom object
instead of tuple. Anyway the ParseResult can be accessed via indexes
as in the old version (python-2.7.10-8.fc23.x86_64)

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>