UBUNTU: SAUCE: drop the interim encoding to utf-8
Generating output via JUnit_api is still failing with the interim step
to encode to 'utf-8'. Futher testing indicates the input is already a
unicode string so encoding it was the wrong step. This time, just
provide the 'backslashreplace' action to replace the unicode character
during the encode to 'ascii'.

Signed-off-by: Francis Ginther <francis.ginther@canonical.com>